NEXUS-5790: Align container and Nx buffer sizes
Preferred way to set response buffer size is to tweak
the container (in case of bundle, Jetty). By default
Jetty uses 32KB buffer, and with this change Nx
will use it too. If this size is inappropriate,
you would want to tweak Jetty embedded in the bundle.

Still, if you need, you can tweak Nx response buffer
size, and Nx will try to pass this information to
the container too.

In bundle case, preferred way is to tweak Jetty,
while the latter mode gives you some freedom
if Nx deployed as WAR in a Servlet container.